// server/src/types/form-profile.ts
export interface FormProfile {
  firstName: string;
  lastName: string;
  fullName: string;
  email: string;
  phone: string;
  gender: 'male' | 'female';
  dateOfBirth: string;
  username: string;
  password: string;
  company: string;
  jobTitle: string;
  address: string;
  city: string;
  state: string;
  zipCode: string;
  country: string;
  url: string;
}